Check empty or nil string before encode
When YF gets HRItemNews and parse it into YFArticle model, string being nil passes along. However, if we want to encode any string properties, this string can not be nil or empty.

I added a check at YFArticle, but it would be nice if `- (NSString *)decodeString:(NSString *)string {` handles it beforehand.